Move junit xml parsing related code to dedicated file
