+++ pkg/report/xml_internal_test.go
